More 64bit update

+++ src/components/objects/hash.c
@@ -125,8 +125,8 @@ SAFFIRE_METHOD(hash, populate) {
         t_object *val = ht_iter_value(&iter);
         ht_iter_next(&iter);
 
-    DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (unsigned int)key, key->ref_count, key->ref_count+1);
-    DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (unsigned int)val, val->ref_count, val->ref_count+1);
+        DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (intptr_t)key, key->ref_count, key->ref_count+1);
+        DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (intptr_t)val, val->ref_count, val->ref_count+1);
         object_inc_ref(key);
         object_inc_ref(val);
         ht_add_obj(self->data.ht, key, val);
@@ -198,8 +198,8 @@ SAFFIRE_METHOD(hash, add) {
         return NULL;
     }
 
-    DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (unsigned int)key, key->ref_count, key->ref_count+1);
-    DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (unsigned int)val, val->ref_count, val->ref_count+1);
+    DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (intptr_t)key, key->ref_count, key->ref_count+1);
+    DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (intptr_t)val, val->ref_count, val->ref_count+1);
 
     object_inc_ref(key);
     object_inc_ref(val);
@@ -383,8 +383,8 @@ static void obj_populate(t_object *obj, t_dll *arg_list) {
         if (! e) break;
         t_object *val = (t_object *)e->data;
 
-    DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (unsigned int)key, key->ref_count, key->ref_count+1);
-    DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (unsigned int)val, val->ref_count, val->ref_count+1);
+    DEBUG_PRINT_CHAR("KEY Hash increasing reference: %08X from %d to %d\n", (intptr_t)key, key->ref_count, key->ref_count+1);
+    DEBUG_PRINT_CHAR("VAL Hash increasing reference: %08X from %d to %d\n", (intptr_t)val, val->ref_count, val->ref_count+1);
 
         object_inc_ref(key);
         object_inc_ref(val);
